+++ upcloud/load_balancer_test.go
@@ -302,8 +302,7 @@ func TestLoadBalancerFrontend(t *testing.T) {
 			"port": 443,
 			"default_backend": "example-backend",
 			"properties": {
-				"timeout_client": 10,
-				"inbound_proxy_protocol": false
+				"timeout_client": 10
 			},
 			"created_at": "2021-12-07T13:58:30.817272Z",
 			"updated_at": "2022-02-11T17:33:08.490581Z"
@@ -350,8 +349,7 @@ func TestLoadBalancerFrontendProperties(t *testing.T) {
 		},
 		`
 		{
-			"timeout_client": 10,
-			"inbound_proxy_protocol": false
+			"timeout_client": 10
 		}
 		`,
 	)
@@ -359,7 +357,7 @@ func TestLoadBalancerFrontendProperties(t *testing.T) {
 		&LoadBalancerFrontendProperties{},
 		&LoadBalancerFrontendProperties{
 			TimeoutClient:        10,
-			InboundProxyProtocol: true,
+			InboundProxyProtocol: BoolPtr(true),
 		},
 		`
 		{
@@ -372,12 +370,42 @@ func TestLoadBalancerFrontendProperties(t *testing.T) {
 		&LoadBalancerFrontendProperties{},
 		&LoadBalancerFrontendProperties{
 			TimeoutClient:        10,
-			InboundProxyProtocol: false,
+			InboundProxyProtocol: BoolPtr(false),
 		},
 		`
 		{
 			"timeout_client": 10,
 			"inbound_proxy_protocol": false
+		}
+		`,
+	)
+	testJSON(t,
+		&LoadBalancerFrontendProperties{},
+		&LoadBalancerFrontendProperties{
+			TimeoutClient:        10,
+			InboundProxyProtocol: BoolPtr(false),
+			HTTP2Enabled:         BoolPtr(false),
+		},
+		`
+		{
+			"timeout_client": 10,
+			"inbound_proxy_protocol": false,
+			"http2_enabled": false
+		}
+		`,
+	)
+	testJSON(t,
+		&LoadBalancerFrontendProperties{},
+		&LoadBalancerFrontendProperties{
+			TimeoutClient:        10,
+			InboundProxyProtocol: BoolPtr(false),
+			HTTP2Enabled:         BoolPtr(true),
+		},
+		`
+		{
+			"timeout_client": 10,
+			"inbound_proxy_protocol": false,
+			"http2_enabled": true
 		}
 		`,
 	)
@@ -570,8 +598,19 @@ func TestLoadBalancerBackendProperties(t *testing.T) {
 	)
 	testJSON(t,
 		&LoadBalancerBackendProperties{},
-		&LoadBalancerBackendProperties{},
-		`{}`,
+		&LoadBalancerBackendProperties{
+			TLSVerify:      BoolPtr(true),
+			TLSEnabled:     BoolPtr(true),
+			TLSUseSystemCA: BoolPtr(true),
+			HTTP2Enabled:   BoolPtr(true),
+		},
+		`{
+			"tls_verify": true,
+			"tls_enabled": true,
+			"tls_use_system_ca": true,
+			"http2_enabled": true
+		}
+		`,
 	)
 }