+++ CHANGELOG.md
@@ -7,6 +7,11 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ### Added
 - Managed Load Balancer: `TLSConfigs` field to `LoadBalancerBackend` to control backend TLS configurations
+- Managed Load Balancer: `TLSEnabled`,  `TLSUseSystemCA`, `TLSVerify` & `HTTP2Enabled` fields to `LoadBalancerBackendProperties`
+- Managed Load Balancer: `HTTP2Enabled` field to `LoadBalancerFrontendProperties`
+
+### Changed
+-  **Breaking**, Managed Load Balancer: `InboundProxyProtocol` field to a boolean pointer in `LoadBalancerFrontendProperties`
 
 ## [6.9.0]
 ### Added